+++ meta_package_manager/tests/test_cli_install.py
@@ -62,6 +62,7 @@ class TestInstall(CLISubCommandTests):
         "steamcmd": "",
         "vscode": "tamasfe.even-better-toml",
         "yarn": "markdown",
+        "zypper": "git",
     }
     assert set(PACKAGE_IDS) == set(pool.all_manager_ids)